+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [4.0.4](https://github.com/factorialco/f0/compare/f0-react-v4.0.3...f0-react-v4.0.4) (2026-06-22)
+
+
+### Bug Fixes
+
+* **SurveyFormBuilder:** foreground input text in dark mode ([#4471](https://github.com/factorialco/f0/issues/4471)) ([17e48d7](https://github.com/factorialco/f0/commit/17e48d70c274f97b2de3cf96e084ebb56d5c707d))
+
+## [4.0.3](https://github.com/factorialco/f0/compare/f0-react-v4.0.2...f0-react-v4.0.3) (2026-06-22)
+
+
+### Bug Fixes
+
+* **F0CardHorizontal:** reserve action button width so the confirm/reject pair never shrinks ([#4525](https://github.com/factorialco/f0/issues/4525)) ([10b0b0b](https://github.com/factorialco/f0/commit/10b0b0b901cf16e8b3e4f45236cf58b60c072b66))
+
 ## [4.0.2](https://github.com/factorialco/f0/compare/f0-react-v4.0.1...f0-react-v4.0.2) (2026-06-19)